PostResponse - count 추가

+++ src/main/java/com/project/mentoridge/modules/board/controller/response/PostResponse.java
@@ -16,9 +16,9 @@ public class PostResponse {
     private String createdAt;
 
     // 좋아요 수
-    private int likingCount = 0;
+    private long likingCount = 0L;
     // 댓글 수
-    private int commentCount = 0;
+    private long commentCount = 0L;
 
     public PostResponse(Post post) {
         this.postId = post.getId();
